Added DOMTest
