+++ package.json
@@ -33,6 +33,7 @@
     "@semantic-release/changelog": "^3.0.6",
     "@semantic-release/git": "^7.0.18",
     "@semantic-release/github": "^5.5.5",
+    "assert": "^2.0.0",
     "jest": "^24.8.0",
     "jsdoc": "^3.6.3",
     "puppeteer": "^2.0.0",
